Move combination nCr tests to a dedicated test file
